Refactor modules discovery to Util class
